+++ src/terminal.js
@@ -6,26 +6,29 @@ const helpers = require('./helpers')
 const TERMINAL_NAME = 'Just Testing'
 const LAST_COMMAND = 'lastCommand'
 
+async function runTerminalCommand (extensionContext, command) {
+  await vscode.workspace.saveAll()
+  extensionContext.workspaceState.update(LAST_COMMAND, command)
+
+  const terminal = obtainTerminal()
+  terminal.show(true)
+  terminal.sendText(command)
+}
+
+function obtainTerminal () {
+  const terminal = vscode.window.terminals.find(terminal => terminal.name === TERMINAL_NAME)
+  if (terminal) return terminal
+
+  return vscode.window.createTerminal(TERMINAL_NAME)
+}
+
 class Command {
   constructor (extensionContext) {
     this.extensionContext = extensionContext
   }
 
   runInTerminal (command) {
-    vscode.workspace.saveAll().then(() => {
-      this.extensionContext.workspaceState.update(LAST_COMMAND, command)
-
-      const terminal = this.obtainTerminal()
-      terminal.show(true)
-      terminal.sendText(command)
-    })
-  }
-
-  obtainTerminal () {
-    const terminal = vscode.window.terminals.find(terminal => terminal.name === TERMINAL_NAME)
-    if (terminal) return terminal
-
-    return vscode.window.createTerminal(TERMINAL_NAME)
+    runTerminalCommand(this.extensionContext, command)
   }
 }
 
@@ -56,4 +59,4 @@ class RunLastCommand extends Command {
   }
 }
 
-module.exports = { TemplateCommand, RunLastCommand }
+module.exports = { TemplateCommand, RunLastCommand, runTerminalCommand }